% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{accident_seatbelt1}
\alias{accident_seatbelt1}
\title{汽车事故与安全带1 (Auto Accidents and Seat Belts I)}
\format{An object of class \code{xtabs} (inherits from \code{table}) of dimension 2 x 2 x 2 x 5.}
\usage{
data(accident_seatbelt1)
}
\description{
汽车事故与安全带1 (Auto Accidents and Seat Belts I)
}
\keyword{datasets}
