% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{accident_seatbelt2}
\alias{accident_seatbelt2}
\title{汽车事故与安全带2 (Auto Accidents and Seat Belts II)}
\format{An object of class \code{xtabs} (inherits from \code{table}) of dimension 2 x 2 x 2 x 2.}
\usage{
data(accident_seatbelt2)
}
\description{
汽车事故与安全带2 (Auto Accidents and Seat Belts II)
}
\keyword{datasets}
