% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{afterlife1}
\alias{afterlife1}
\title{是否相信来世1 (Belief in Afterlife I)}
\format{An object of class \code{xtabs} (inherits from \code{table}) with 2 rows and 2 columns.}
\usage{
data(afterlife1)
}
\description{
是否相信来世1 (Belief in Afterlife I)
}
\keyword{datasets}
