% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{afterlife2}
\alias{afterlife2}
\title{是否相信来世2 (Belief in Afterlife II)}
\format{An object of class \code{xtabs} (inherits from \code{table}) of dimension 2 x 2 x 3.}
\usage{
data(afterlife2)
}
\description{
是否相信来世2 (Belief in Afterlife II)
}
\keyword{datasets}
