% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AIDS_treatment}
\alias{AIDS_treatment}
\title{对AIDS的处理方法 (Treatment of AIDS)}
\format{An object of class \code{xtabs} (inherits from \code{table}) of dimension 2 x 2 x 2.}
\usage{
data(AIDS_treatment)
}
\description{
对AIDS的处理方法 (Treatment of AIDS)
}
\keyword{datasets}
