% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{albumin}
\alias{albumin}
\title{控制蛋白质与死亡率 (Albumin Administering and Mortality)}
\format{An object of class \code{data.frame} with 32 rows and 8 columns.}
\usage{
data(albumin)
}
\description{
控制蛋白质与死亡率 (Albumin Administering and Mortality)
}
\keyword{datasets}
