% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{alligators1}
\alias{alligators1}
\title{短吻鳄食物选择1 (Alligator Food Choice I)}
\format{An object of class \code{data.frame} with 59 rows and 2 columns.}
\usage{
data(alligators1)
}
\description{
短吻鳄食物选择1 (Alligator Food Choice I)
}
\keyword{datasets}
