% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{alligators2}
\alias{alligators2}
\title{短吻鳄食物选择2 (Alligator Food Choice II)}
\format{An object of class \code{xtabs} (inherits from \code{table}) of dimension 2 x 4 x 2 x 5.}
\usage{
data(alligators2)
}
\description{
短吻鳄食物选择2 (Alligator Food Choice II)
}
\keyword{datasets}
