% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aspirin}
\alias{aspirin}
\title{阿司匹林与心脏病 (Aspirin and Heart Attacks)}
\format{An object of class \code{xtabs} (inherits from \code{table}) with 2 rows and 2 columns.}
\usage{
data(aspirin)
}
\description{
阿司匹林与心脏病 (Aspirin and Heart Attacks)
}
\keyword{datasets}
