% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{athlete_graduate}
\alias{athlete_graduate}
\title{新生运动员毕业比例 (Graduation Rates of Freshman Athletes)}
\format{An object of class \code{data.frame} with 4 rows and 3 columns.}
\usage{
data(athlete_graduate)
}
\description{
新生运动员毕业比例 (Graduation Rates of Freshman Athletes)
}
\keyword{datasets}
