% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AZT}
\alias{AZT}
\title{AZT的使用与AIDS (AZT Use and AIDS)}
\format{An object of class \code{xtabs} (inherits from \code{table}) of dimension 2 x 2 x 2.}
\usage{
data(AZT)
}
\description{
AZT的使用与AIDS (AZT Use and AIDS)
}
\keyword{datasets}
