% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{birth_control}
\alias{birth_control}
\title{是否应该给青少年提供避孕方法 (Whether Methods of Birth Control Should Be Available to Teenagers)}
\format{An object of class \code{xtabs} (inherits from \code{table}) with 9 rows and 4 columns.}
\usage{
data(birth_control)
}
\description{
是否应该给青少年提供避孕方法 (Whether Methods of Birth Control Should Be Available to Teenagers)
}
\keyword{datasets}
