% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{blood_pressure}
\alias{blood_pressure}
\title{心脏病与血压的关系 (Heart Disease and Blood Pressure)}
\format{An object of class \code{data.frame} with 8 rows and 3 columns.}
\usage{
data(blood_pressure)
}
\description{
心脏病与血压的关系 (Heart Disease and Blood Pressure)
}
\keyword{datasets}
