% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cancer_remission}
\alias{cancer_remission}
\title{癌症缓解 (Cancer Remission)}
\format{An object of class \code{data.frame} with 14 rows and 3 columns.}
\usage{
data(cancer_remission)
}
\description{
癌症缓解 (Cancer Remission)
}
\keyword{datasets}
