% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdabookdb-package.R
\docType{package}
\name{cdabookdb}
\alias{cdabookdb}
\alias{cdabookdb-package}
\title{DataBase for  "An Introduction to Categorical Data Analysis, 2nd edition by Alan Agresti"}
\description{
DataBase for  "An Introduction to Categorical Data Analysis, 2nd edition by Alan Agresti"

  Some of data files in this packages are from package "icda" by Brett Presnell
  (http://users.stat.ufl.edu/~presnell/Courses/sta4504-2012sp/R-info.shtml).
}
\details{
cdabookdb


This package include data sets that may be used in the textbook.
Using \code{data(package = "cdabookdb")} to see all the data sets
that are avaliable in this package.
}
