% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chip_imperfection}
\alias{chip_imperfection}
\title{晶片瑕疵 (Chip Imperfection)}
\format{An object of class \code{data.frame} with 20 rows and 3 columns.}
\usage{
data(chip_imperfection)
}
\description{
晶片瑕疵 (Chip Imperfection)
}
\keyword{datasets}
