% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cholesterol}
\alias{cholesterol}
\title{LDL胆固醇研究 (LDL Cholesterol Study)}
\format{An object of class \code{xtabs} (inherits from \code{table}) of dimension 4 x 4 x 2.}
\usage{
data(cholesterol)
}
\description{
LDL胆固醇研究 (LDL Cholesterol Study)
}
\keyword{datasets}
