% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{credit_score}
\alias{credit_score}
\title{信用评分 (Credit Score)}
\format{An object of class \code{data.frame} with 1000 rows and 8 columns.}
\usage{
data(credit_score)
}
\description{
来自Fahrmeir包中的credit数据集，可查看相应的帮助页面以获取变量信息
}
\keyword{datasets}
