% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{creditcard}
\alias{creditcard}
\title{收入与是否拥有旅行信用卡 (Relation Between Income and Whether One Possesses a Travel Credit Card)}
\format{An object of class \code{data.frame} with 31 rows and 3 columns.}
\usage{
data(creditcard)
}
\description{
收入与是否拥有旅行信用卡 (Relation Between Income and Whether One Possesses a Travel Credit Card)
}
\keyword{datasets}
