% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{deathpenalty2}
\alias{deathpenalty2}
\title{死刑判决2 (Death Penalty II)}
\format{An object of class \code{xtabs} (inherits from \code{table}) of dimension 2 x 2 x 2.}
\usage{
data(deathpenalty2)
}
\description{
死刑判决2 (Death Penalty II)
}
\keyword{datasets}
