% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{edu_aspiration}
\alias{edu_aspiration}
\title{高中生教育期望与家庭收入 (Educational Aspirations of High School Students and Family Income)}
\format{An object of class \code{xtabs} (inherits from \code{table}) with 4 rows and 3 columns.}
\usage{
data(edu_aspiration)
}
\description{
高中生教育期望与家庭收入 (Educational Aspirations of High School Students and Family Income)
}
\keyword{datasets}
