% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{environmental_protection}
\alias{environmental_protection}
\title{为了环保是否愿意付更高的税或降低生活质量 (Is It Willing to Pay Higher Taxes or Cut Living Standards for Environmental Protection?)}
\format{An object of class \code{xtabs} (inherits from \code{table}) with 2 rows and 2 columns.}
\usage{
data(environmental_protection)
}
\description{
为了环保是否愿意付更高的税或降低生活质量 (Is It Willing to Pay Higher Taxes or Cut Living Standards for Environmental Protection?)
}
\keyword{datasets}
