% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{football_arrest}
\alias{football_arrest}
\title{足球联赛观赛人数与被捕人数 (Total Attendance and Total Number of Arrests in Football League)}
\format{An object of class \code{data.frame} with 23 rows and 3 columns.}
\usage{
data(football_arrest)
}
\description{
足球联赛观赛人数与被捕人数 (Total Attendance and Total Number of Arrests in Football League)
}
\keyword{datasets}
