% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gender_party}
\alias{gender_party}
\title{政党认同中的性别差异 (Gender Gap in Political Affiliation)}
\format{An object of class \code{xtabs} (inherits from \code{table}) with 2 rows and 3 columns.}
\usage{
data(gender_party)
}
\description{
政党认同中的性别差异 (Gender Gap in Political Affiliation)
}
\keyword{datasets}
