% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{government_spending}
\alias{government_spending}
\title{政府对环境、卫生、大城市补助、法律强制执行的拨款问题 (Government Spending on the Environment, Health, Assistance to Big Cities, and Law Enforcement)}
\format{An object of class \code{xtabs} (inherits from \code{table}) of dimension 3 x 3 x 3 x 3.}
\usage{
data(government_spending)
}
\description{
政府对环境、卫生、大城市补助、法律强制执行的拨款问题 (Government Spending on the Environment, Health, Assistance to Big Cities, and Law Enforcement)
}
\keyword{datasets}
