% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{happiness1}
\alias{happiness1}
\title{家庭收入与幸福感 (A Person’s Perceived Happiness with Their Family Income)}
\format{An object of class \code{xtabs} (inherits from \code{table}) with 3 rows and 3 columns.}
\usage{
data(happiness1)
}
\description{
家庭收入与幸福感 (A Person’s Perceived Happiness with Their Family Income)
}
\keyword{datasets}
