% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{happiness2}
\alias{happiness2}
\title{幸福与宗教参与 (Happiness and Religion Attendance)}
\format{An object of class \code{xtabs} (inherits from \code{table}) with 3 rows and 3 columns.}
\usage{
data(happiness2)
}
\description{
幸福与宗教参与 (Happiness and Religion Attendance)
}
\keyword{datasets}
