% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{happiness3}
\alias{happiness3}
\title{幸福感与婚姻状态 (Happiness and Marital Status)}
\format{An object of class \code{xtabs} (inherits from \code{table}) with 2 rows and 3 columns.}
\usage{
data(happiness3)
}
\description{
幸福感与婚姻状态 (Happiness and Marital Status)
}
\keyword{datasets}
