% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{horseshoecrabs}
\alias{horseshoecrabs}
\title{母鲎数据 (Female Horseshoe Crabs)}
\format{An object of class \code{data.frame} with 173 rows and 5 columns.}
\usage{
data(horseshoecrabs)
}
\description{
母鲎数据 (Female Horseshoe Crabs)
}
\keyword{datasets}
