% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ideology}
\alias{ideology}
\title{政治意识形态、性别与隶属党派 (Political Ideology, Gender and Party Affiliation)}
\format{An object of class \code{xtabs} (inherits from \code{table}) of dimension 2 x 2 x 5.}
\usage{
data(ideology)
}
\description{
政治意识形态、性别与隶属党派 (Political Ideology, Gender and Party Affiliation)
}
\keyword{datasets}
