% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{impairment}
\alias{impairment}
\title{心理健康、SES与生活事件 (Mental Impairment, SES and Life Events)}
\format{An object of class \code{data.frame} with 40 rows and 4 columns.}
\usage{
data(impairment)
}
\description{
心理健康、SES与生活事件 (Mental Impairment, SES and Life Events)
}
\keyword{datasets}
