% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{incontinent}
\alias{incontinent}
\title{个人使用药物后是否失禁 (Whether a Subject Became Incontinent After Received a Drug)}
\format{An object of class \code{data.frame} with 21 rows and 4 columns.}
\usage{
data(incontinent)
}
\description{
个人使用药物后是否失禁 (Whether a Subject Became Incontinent After Received a Drug)
}
\keyword{datasets}
