% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{job_satisfaction1}
\alias{job_satisfaction1}
\title{工作满意度调查1 (Survey on Job Satisfaction I)}
\format{An object of class \code{xtabs} (inherits from \code{table}) of dimension 2 x 3 x 7 x 2 x 2.}
\usage{
data(job_satisfaction1)
}
\description{
工作满意度调查1 (Survey on Job Satisfaction I)
}
\keyword{datasets}
