% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{job_satisfaction2}
\alias{job_satisfaction2}
\title{工作满意度调查2 (Survey on Job Satisfaction II)}
\format{An object of class \code{xtabs} (inherits from \code{table}) of dimension 2 x 4 x 4.}
\usage{
data(job_satisfaction2)
}
\description{
工作满意度调查2 (Survey on Job Satisfaction II)
}
\keyword{datasets}
