% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kyphosis_age}
\alias{kyphosis_age}
\title{是否出现驼背与以月计算的年龄 (Whether Kyphosis is Present and the Age in Months)}
\format{An object of class \code{data.frame} with 40 rows and 2 columns.}
\usage{
data(kyphosis_age)
}
\description{
是否出现驼背与以月计算的年龄 (Whether Kyphosis is Present and the Age in Months)
}
\keyword{datasets}
