% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{larynx_cancer}
\alias{larynx_cancer}
\title{喉癌治疗中手术与放疗的对比 (Comparing Radiation Therapy with Surgery in Treating Cancer of the Larynx)}
\format{An object of class \code{xtabs} (inherits from \code{table}) with 2 rows and 2 columns.}
\usage{
data(larynx_cancer)
}
\description{
喉癌治疗中手术与放疗的对比 (Comparing Radiation Therapy with Surgery in Treating Cancer of the Larynx)
}
\keyword{datasets}
