% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lungcancer_treatment}
\alias{lungcancer_treatment}
\title{小细胞肺癌治疗 (Small-cell Lung Cancer Treatment)}
\format{An object of class \code{xtabs} (inherits from \code{table}) of dimension 2 x 2 x 4.}
\usage{
data(lungcancer_treatment)
}
\description{
小细胞肺癌治疗 (Small-cell Lung Cancer Treatment)
}
\keyword{datasets}
