% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{malformation}
\alias{malformation}
\title{饮酒与胎儿畸形 (Alcohol Use and Infant Malformation)}
\format{An object of class \code{xtabs} (inherits from \code{table}) with 5 rows and 2 columns.}
\usage{
data(malformation)
}
\description{
饮酒与胎儿畸形 (Alcohol Use and Infant Malformation)
}
\keyword{datasets}
