% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{marijuana}
\alias{marijuana}
\title{高中毕业生的酒、香烟、大麻的使用情况 (Alcohol, Cigarette, and Marijuana Use for High School Seniors)}
\format{An object of class \code{table} of dimension 2 x 2 x 2 x 2 x 2.}
\usage{
data(marijuana)
}
\description{
高中毕业生的酒、香烟、大麻的使用情况 (Alcohol, Cigarette, and Marijuana Use for High School Seniors)
}
\keyword{datasets}
