% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{marital_happiness}
\alias{marital_happiness}
\title{婚姻幸福与家庭收入 (Marital Happiness and Family Income)}
\format{An object of class \code{xtabs} (inherits from \code{table}) with 3 rows and 3 columns.}
\usage{
data(marital_happiness)
}
\description{
婚姻幸福与家庭收入 (Marital Happiness and Family Income)
}
\keyword{datasets}
