% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MBtest1}
\alias{MBtest1}
\title{Myers-Briggs人格测验1 (Myers-Briggs Personality Test I)}
\format{An object of class \code{xtabs} (inherits from \code{table}) of dimension 2 x 2 x 2 x 2 x 2.}
\usage{
data(MBtest1)
}
\description{
Myers-Briggs人格测验1 (Myers-Briggs Personality Test I)
}
\keyword{datasets}
