% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MBtest2}
\alias{MBtest2}
\title{Myers-Briggs人格测验2 (Myers-Briggs Personality Test II)}
\format{An object of class \code{xtabs} (inherits from \code{table}) of dimension 2 x 2 x 2 x 2 x 2.}
\usage{
data(MBtest2)
}
\description{
Myers-Briggs人格测验2 (Myers-Briggs Personality Test II)
}
\keyword{datasets}
