% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MBtest3}
\alias{MBtest3}
\title{Myers-Briggs人格测验3 (Myers-Briggs Personality Test III)}
\format{An object of class \code{xtabs} (inherits from \code{table}) of dimension 2 x 2 x 2 x 2.}
\usage{
data(MBtest3)
}
\description{
Myers-Briggs人格测验3 (Myers-Briggs Personality Test III)
}
\keyword{datasets}
