% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{merit_pay_race}
\alias{merit_pay_race}
\title{绩效工资与种族 (Merit Pay and Race)}
\format{An object of class \code{xtabs} (inherits from \code{table}) of dimension 5 x 2 x 2.}
\usage{
data(merit_pay_race)
}
\description{
绩效工资与种族 (Merit Pay and Race)
}
\keyword{datasets}
