% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{missing_persons}
\alias{missing_persons}
\title{伦敦失踪人口 (Missing Persons in London)}
\format{An object of class \code{data.frame} with 6 rows and 4 columns.}
\usage{
data(missing_persons)
}
\description{
伦敦失踪人口 (Missing Persons in London)
}
\keyword{datasets}
