% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{osteosarcoma}
\alias{osteosarcoma}
\title{非转移骨肉瘤 (Nonmetastatic Osteosarcoma)}
\format{An object of class \code{xtabs} (inherits from \code{table}) of dimension 2 x 2 x 2 x 2.}
\usage{
data(osteosarcoma)
}
\description{
非转移骨肉瘤 (Nonmetastatic Osteosarcoma)
}
\keyword{datasets}
