% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{premarital_sex1}
\alias{premarital_sex1}
\title{婚前性行为1 (Premarital Sex I)}
\format{An object of class \code{xtabs} (inherits from \code{table}) of dimension 3 x 2 x 2 x 2.}
\usage{
data(premarital_sex1)
}
\description{
婚前性行为1 (Premarital Sex I)
}
\keyword{datasets}
