% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{premarital_sex2}
\alias{premarital_sex2}
\title{婚前性行为2 (Premarital Sex II)}
\format{An object of class \code{xtabs} (inherits from \code{table}) with 4 rows and 4 columns.}
\usage{
data(premarital_sex2)
}
\description{
婚前性行为2 (Premarital Sex II)
}
\keyword{datasets}
