% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{promotion_race}
\alias{promotion_race}
\title{晋升能力与种族 (Promotion Ability and Race)}
\format{An object of class \code{xtabs} (inherits from \code{table}) of dimension 2 x 3 x 2.}
\usage{
data(promotion_race)
}
\description{
晋升能力与种族 (Promotion Ability and Race)
}
\keyword{datasets}
