% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{psych_diag_drugs}
\alias{psych_diag_drugs}
\title{精神病诊断与药物处方 (Psychiatric Diagnosis and Drugs Prescribed)}
\format{An object of class \code{xtabs} (inherits from \code{table}) with 5 rows and 2 columns.}
\usage{
data(psych_diag_drugs)
}
\description{
精神病诊断与药物处方 (Psychiatric Diagnosis and Drugs Prescribed)
}
\keyword{datasets}
