% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rabbit_penicillin}
\alias{rabbit_penicillin}
\title{兔子与青霉素 (Rabbits and Penicillin)}
\format{An object of class \code{data.frame} with 10 rows and 4 columns.}
\usage{
data(rabbit_penicillin)
}
\description{
兔子与青霉素 (Rabbits and Penicillin)
}
\keyword{datasets}
