% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{race_party}
\alias{race_party}
\title{政党认同与种族 (Party Identification and Race)}
\format{An object of class \code{xtabs} (inherits from \code{table}) with 2 rows and 3 columns.}
\usage{
data(race_party)
}
\description{
政党认同与种族 (Party Identification and Race)
}
\keyword{datasets}
