% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{religious_belief}
\alias{religious_belief}
\title{受教育程度与宗教信仰 (Highest Degree of Education and Religious Beliefs)}
\format{An object of class \code{xtabs} (inherits from \code{table}) with 3 rows and 3 columns.}
\usage{
data(religious_belief)
}
\description{
受教育程度与宗教信仰 (Highest Degree of Education and Religious Beliefs)
}
\keyword{datasets}
