% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{smoking_cd}
\alias{smoking_cd}
\title{吸烟与冠心病死亡 (Smoking and Coronary Deaths)}
\format{An object of class \code{xtabs} (inherits from \code{table}) of dimension 5 x 2 x 2.}
\usage{
data(smoking_cd)
}
\description{
吸烟与冠心病死亡 (Smoking and Coronary Deaths)
}
\keyword{datasets}
