% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{smoking_lungcancer}
\alias{smoking_lungcancer}
\title{肺癌与吸烟之间的关系 (Link Between Lung Cancer and Smoking)}
\format{An object of class \code{xtabs} (inherits from \code{table}) with 2 rows and 2 columns.}
\usage{
data(smoking_lungcancer)
}
\description{
肺癌与吸烟之间的关系 (Link Between Lung Cancer and Smoking)
}
\keyword{datasets}
