% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{smoking_lungcancer_cn}
\alias{smoking_lungcancer_cn}
\title{中国的吸烟与肺癌数据 (Smoking and Lung Cancer in China)}
\format{An object of class \code{table} of dimension 8 x 2 x 2.}
\usage{
data(smoking_lungcancer_cn)
}
\description{
中国的吸烟与肺癌数据 (Smoking and Lung Cancer in China)
}
\keyword{datasets}
