% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{smoking_mi}
\alias{smoking_mi}
\title{吸烟状态与心肌梗死 (Smoking Status and Myocardial Infarction)}
\format{An object of class \code{xtabs} (inherits from \code{table}) with 2 rows and 2 columns.}
\usage{
data(smoking_mi)
}
\description{
吸烟状态与心肌梗死 (Smoking Status and Myocardial Infarction)
}
\keyword{datasets}
