% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{snoring_heartdisease}
\alias{snoring_heartdisease}
\title{打鼾与心脏病 (Snoring and Heart Disease)}
\format{An object of class \code{xtabs} (inherits from \code{table}) with 4 rows and 2 columns.}
\usage{
data(snoring_heartdisease)
}
\description{
打鼾与心脏病 (Snoring and Heart Disease)
}
\keyword{datasets}
