% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{teen_sex}
\alias{teen_sex}
\title{青少年与性行为 (Teenagers and Sexual Intercourse)}
\format{An object of class \code{xtabs} (inherits from \code{table}) of dimension 2 x 2 x 2.}
\usage{
data(teen_sex)
}
\description{
青少年与性行为 (Teenagers and Sexual Intercourse)
}
\keyword{datasets}
