% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{teenager_crime}
\alias{teenager_crime}
\title{使得青少年犯罪增加的因素 (Factors that Increase Teenager Crime)}
\format{An object of class \code{xtabs} (inherits from \code{table}) with 2 rows and 3 columns.}
\usage{
data(teenager_crime)
}
\description{
使得青少年犯罪增加的因素 (Factors that Increase Teenager Crime)
}
\keyword{datasets}
