% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{temperature_distress}
\alias{temperature_distress}
\title{飞行温度与是否受热遇险 (Temperature and Whether at Least One Primary O-ring Suffered Thermal Distress)}
\format{An object of class \code{data.frame} with 23 rows and 3 columns.}
\usage{
data(temperature_distress)
}
\description{
飞行温度与是否受热遇险 (Temperature and Whether at Least One Primary O-ring Suffered Thermal Distress)
}
\keyword{datasets}
