% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{throat}
\alias{throat}
\title{手术后喉咙痛 (Sore Throat after Surgery)}
\format{An object of class \code{data.frame} with 35 rows and 3 columns.}
\usage{
data(throat)
}
\description{
手术后喉咙痛 (Sore Throat after Surgery)
}
\keyword{datasets}
