% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{toxicity}
\alias{toxicity}
\title{怀孕老鼠与发育毒性研究 (Pregnant Mice and Developmental Toxicity Study)}
\format{An object of class \code{xtabs} (inherits from \code{table}) with 5 rows and 3 columns.}
\usage{
data(toxicity)
}
\description{
怀孕老鼠与发育毒性研究 (Pregnant Mice and Developmental Toxicity Study)
}
\keyword{datasets}
