% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{traincollisions}
\alias{traincollisions}
\title{英国的火车事故 (British Train Accidents over Time)}
\format{An object of class \code{data.frame} with 29 rows and 4 columns.}
\usage{
data(traincollisions)
}
\description{
英国的火车事故 (British Train Accidents over Time)
}
\keyword{datasets}
