% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{treatment2}
\alias{treatment2}
\title{治疗中心、治疗方案与响应2 (Clinic, Treatment and Response II)}
\format{An object of class \code{xtabs} (inherits from \code{table}) of dimension 8 x 2 x 2.}
\usage{
data(treatment2)
}
\description{
治疗中心、治疗方案与响应2 (Clinic, Treatment and Response II)
}
\keyword{datasets}
