% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{treatment3}
\alias{treatment3}
\title{治疗中心、治疗方案与响应3 (Clinic, Treatment and Response III)}
\format{An object of class \code{xtabs} (inherits from \code{table}) of dimension 5 x 2 x 2.}
\usage{
data(treatment3)
}
\description{
治疗中心、治疗方案与响应3 (Clinic, Treatment and Response III)
}
\keyword{datasets}
