% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{UCBAdmissions}
\alias{UCBAdmissions}
\title{加州大学伯克利分校研究生录取 (Graduate Admissions at the University of California, Berkeley)}
\format{An object of class \code{table} of dimension 2 x 2 x 6.}
\usage{
UCBAdmissions
}
\description{
来自datasets包中的UCBAdmissions数据集，可查看相应的帮助页面以获取变量信息
}
\keyword{datasets}
