% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{UFAdmissions}
\alias{UFAdmissions}
\title{佛罗里达大学的研究生入学 (Graduate Admissions at the University of Florida)}
\format{An object of class \code{xtabs} (inherits from \code{table}) of dimension 23 x 2 x 2.}
\usage{
data(UFAdmissions)
}
\description{
佛罗里达大学的研究生入学 (Graduate Admissions at the University of Florida)
}
\keyword{datasets}
