% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{white_black_acceptance}
\alias{white_black_acceptance}
\title{白人对于黑人的接受度 (The Acceptance of Blacks by Whites)}
\format{An object of class \code{xtabs} (inherits from \code{table}) of dimension 2 x 2 x 2.}
\usage{
data(white_black_acceptance)
}
\description{
白人对于黑人的接受度 (The Acceptance of Blacks by Whites)
}
\keyword{datasets}
